BSOL Theme Colours

This function returns the colours that are used in the BSOL ICS theme as a named vector of RGB values in hexadecimal form.

Usage

bsol_theme_cols(
  ...,
  palette = c(NA, "main", "alternative", "ics_blue", "ics_orange", "ics_green",
    "ics_purple", "ics_navy")
)

Arguments

...

individual colours that you wish to get the values of

palette

a name of a palette to select the colours of

Value

a named vector of RGB colours in hexadecimal form

Details

If no arguments are passed to the function, then all of the colours are returned. If only specific colours are required, then the names of the colours that are required can be passed as strings to the function, and only those colours will be returned.

If a "palette" is selected then no arguments should be passed to ...

Only one palette can be selected at a time.

Examples

# show all of the colours
bsol_theme_cols()
#>       green  light_blue      orange   deep_navy      purple    nhs_blue 
#>   "#8cedab"   "#4fbff0"   "#fc8700"   "#031d44"   "#b88ce3"   "#005EB8" 
#> light_slate    charcoal       white 
#>   "#b2b7b9"   "#2c2825"   "#ffffff" 
# or, just show some colours
bsol_theme_cols("green", "orange", "deep_navy", "nhs_blue")
#>     green    orange deep_navy  nhs_blue 
#> "#8cedab" "#fc8700" "#031d44" "#005EB8" 
# or, select a single palette
bsol_theme_cols(palette = "ics_orange")
#>    orange     white deep_navy 
#> "#fc8700" "#ffffff" "#031d44"