ggplotly BSOL wrapper for passing the font size properly

ggplotly BSOL wrapper for passing the font size properly

Usage

ggplotly_bsol(p = ggplot2::last_plot(), font, ...)

Arguments

p

a ggplot object, default is the last plot called

font

required font, extracted from the plot if not specified otherwise

...

Arguments to the layout object. For documentation, see https://plotly.com/r/reference/#Layout_and_layout_style_objects

Value

A plotly object

Examples

library(ggplot2)
a<-ggplot(iris, aes(y=Sepal.Length, x=Sepal.Width, colour=Species))+
geom_point()+
scale_colour_bsol()+
theme_bsol()

ggplotly_bsol(a)